% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_eco_data}
\alias{get_eco_data}
\title{Récupérer les données d'un seul compteur}
\usage{
get_eco_data(
  station,
  flowIds = NULL,
  debut,
  fin,
  interval = "4",
  idOrganisme = NULL
)
}
\arguments{
\item{station}{Une ligne (data.frame) issue de get_stations() OU l'ID du point de comptage.}

\item{flowIds}{Facultatif si 'station' est fourni. Chaîne des flux (ex: "123;456").}

\item{debut}{Date de début (JJ/MM/AAAA).}

\item{fin}{Date de fin (JJ/MM/AAAA).}

\item{interval}{Code intervalle (5: Semaine, 4: Jour (default), 3: Heure).}

\item{idOrganisme}{Facultatif si 'station' est fourni. ID de l'organisme.}
}
\description{
Récupérer les données d'un seul compteur
}
