% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_multi_eco_data}
\alias{get_multi_eco_data}
\title{Récupérer les données de plusieurs compteurs (Bulk)}
\usage{
get_multi_eco_data(stations_df, debut, fin, interval = "4")
}
\arguments{
\item{stations_df}{Le data.frame généré par get_stations().}

\item{debut}{Date de début (JJ/MM/AAAA).}

\item{fin}{Date de fin (JJ/MM/AAAA).}

\item{interval}{Code intervalle (5: Semaine, 4: Jour (default), 3: Heure).}
}
\description{
Récupérer les données de plusieurs compteurs (Bulk)
}
