% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Récupérer les stations d'un organisme}
\usage{
get_stations(idOrganisme, code_pratique = NULL)
}
\arguments{
\item{idOrganisme}{ID de l'organisme (ex: "3902" pour Lyon, "857" pour Düsseldorf...)}

\item{code_pratique}{Facultatif. Un entier (ex: 2 pour vélo) pour filtrer un type spécifique.
Si NULL (par défaut), toutes les stations sont retournées.}
}
\description{
Récupérer les stations d'un organisme
}
